pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "capiscio-mcp"
version = "0.1.0"
description = "Trust badges for MCP tool calls - RFC-006 & RFC-007 implementation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "CapiscIO Team", email = "engineering@capisc.io" }
]
keywords = ["mcp", "trust", "security", "ai", "agents", "badges", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "requests>=2.31.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0",
]
crypto = [
    "cryptography>=42.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "cryptography>=42.0.0",
]

[project.urls]
Homepage = "https://capisc.io"
Documentation = "https://docs.capisc.io/mcp"
Repository = "https://github.com/capiscio/capiscio-mcp-python"
Issues = "https://github.com/capiscio/capiscio-mcp-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["capiscio_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["capiscio_mcp"]
omit = [
    "capiscio_mcp/_proto/*",
    "*/__pycache__/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
fail_under = 80
show_missing = true
skip_covered = false

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    "capiscio_mcp/_proto/*",
]

[tool.ruff.lint]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    "capiscio_mcp/_proto/",
]

[[tool.mypy.overrides]]
module = [
    "grpc",
    "grpc.*",
    "google.protobuf.*",
    "mcp.*",
]
ignore_missing_imports = true
